+++ packages/translation-nmtcpp/addonLogging.d.ts
@@ -4,4 +4,6 @@ export interface AddonLogging {
     releaseLogger: () => void;
 }
 declare const addonLogging: AddonLogging;
+export declare const setLogger: (callback: NativeLoggerCallback) => void;
+export declare const releaseLogger: () => void;
 export default addonLogging;